Clean up `InstanceMut` access
Previous `InstanceWorldMut` was used as a temporary solution to handle the difference between `EntityMut` and `EntityWorldMut` and associated lifetime complexities.

But `InstanceWorldMut` didn't make sense as implemented and didn't match the rest of the API. For example, the `Deref` impl on it would give `EntityWorldMut` instead of `&mut T`, unlike `InstanceMut`.

This change just removes `InstanceWorldMut` and instead solves the underlying lifetime issues to bring `InstanceMut` closer to `(Instance<T>, &mut T)` when accessed from world.